Modify: index.js split to module
