+++ taipei-day-trip/function/common/server.js
@@ -1,0 +1,6 @@
+// Get the current domain
+let currentDomain = window.location.hostname;
+// Get the current port
+let currentPort = window.location.port;
+// Combine domain and port
+export let server = `http://${currentDomain}:${currentPort}`
