+++ taipei-day-trip/static/attraction.html
@@ -6,7 +6,7 @@
   <title>Attraction</title>
   <link rel="stylesheet" href="/style/style.css">
 </head>
-<script defer src="/function/attraction.js"></script>
+<script defer type="module" src="/function/attraction.js"></script>
 <body>
   <div class="navbar">
     <div class="navbar--title">台北一日遊</div>
