+++ taipei-day-trip/static/index.html
@@ -6,7 +6,7 @@
   <title>Index</title>
   <link rel="stylesheet" href="/style/style.css">
 </head>
-<script defer src="/function/index.js"></script>
+<script defer type="module" src="/function/index.js"></script>
 <body>
   <div class="navbar">
     <div class="navbar--title">台北一日遊</div>
